requirements.txt
numpy
pandas
matplotlib
geopandas
shapely
pyproj
contextily
ffmpeg
paramiko
python-dotenv
scp
